suicide

+++ nwg_clipman/main.py
@@ -64,15 +64,13 @@ def terminate_old_instance():
     if os.path.isfile(pid_file):
         try:
             old_pid = int(load_text_file(pid_file))
-            # if old_pid != pid:
-            eprint(f"Attempting to kill the old instance in case it's still running, pid: {old_pid}")
-            os.kill(old_pid, 15)
-            print("terminating")
-            sys.exit(0)
+            if old_pid != pid:
+                eprint(f"Attempting to kill the old instance in case it's still running, pid: {old_pid}")
+                os.kill(old_pid, 15)
         except:
             pass
     # save new pid
-    save_string(str(os.getpid()), pid_file)
+    save_string(str(pid), pid_file)
 
 
 def handle_keyboard(win, event):
